@@ -10,27 +10,42 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "axios": "^1.7.3",
+    "@ckeditor/ckeditor5-build-classic": "^43.0.0",
+    "@ckeditor/ckeditor5-react": "^9.0.0",
+    "@paypal/react-paypal-js": "^8.5.0",
+    "autoprefixer": "^10.4.20",
+    "axios": "^1.7.4",
+    "chart.js": "^4.4.3",
     "dayjs": "^1.11.12",
     "js-cookie": "^3.0.5",
     "jwt-decode": "^4.0.0",
+    "lucide-react": "^0.428.0",
+    "moment": "^2.30.1",
+    "postcss": "^8.4.41",
     "react": "^18.3.1",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "^18.3.1",
-    "react-router-dom": "^6.26.0",
-    "tailwind": "^4.0.0",
-    "zustand": "^4.5.4"
+    "react-hook-form": "^7.52.2",
+    "react-icons": "^5.3.0",
+    "react-photo-album": "^3.0.1",
+    "react-router-dom": "^6.26.1",
+    "sweetalert2": "^11.12.4",
+    "tailwindcss": "^3.4.10",
+    "yet-another-react-lightbox": "^3.21.3",
+    "zustand": "^4.5.5"
   },
   "devDependencies": {
-    "@eslint/js": "^9.8.0",
+    "@eslint/js": "^9.9.0",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
     "@vitejs/plugin-react": "^4.3.1",
-    "eslint": "^9.8.0",
+    "eslint": "^9.9.0",
     "eslint-plugin-react": "^7.35.0",
-    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-react-refresh": "^0.4.9",
     "globals": "^15.9.0",
-    "tailwindcss": "^3.4.9",
-    "vite": "^5.4.0"
+    "prettier": "^3.3.3",
+    "simple-zustand-devtools": "^1.1.0",
+    "vite": "^5.4.1"
   }
 }
