@@ -1,35 +1,36 @@
 import { useState } from 'react'
-import reactLogo from './assets/react.svg'
-import viteLogo from '/vite.svg'
+import { BrowserRouter as Router, Route, Routes, useLocation } from 'react-router-dom'
+
+import LoginPage from './views/auth/LoginPage';
+import RegisterPage from './views/auth/RegisterPage';
 
 
-function App() {
-  const [count, setCount] = useState(0)
-
+const App = () => {
   return (
     <>
-      <div>
-        <a href="https://vitejs.dev" target="_blank">
-          <img src={viteLogo} className="logo" alt="Vite logo" />
-        </a>
-        <a href="https://react.dev" target="_blank">
-          <img src={reactLogo} className="logo react" alt="React logo" />
-        </a>
-      </div>
-      <h1>Vite + React</h1>
-      <div className="card">
-        <button onClick={() => setCount((count) => count + 1)}>
-          count is {count}
-        </button>
-        <p>
-          Edit <code>src/App.jsx</code> and save to test HMR
-        </p>
-      </div>
-      <p className="read-the-docs">
-        Click on the Vite and React logos to learn more
-      </p>
+    <Router>
+      <AppContent />
+    </Router>
     </>
-  )
-}
+  );
+};
+
+const AppContent = () => {
+  const location = useLocation();
+  const hideNavbarRoutes = ['/login', 'register'];
+  const isNavbarVisible = !hideNavbarRoutes.includes(location.pathname);
+
+  return (
+    <div className="flex flex-col h-screen overflow-hidden">
+      {/*isNavbarVisible && <Sidebar />*/}
+      <div className="flex-grow overflow-hidden">
+        <Routes>
+          <Route path="/login" element={<LoginPage />} />
+          <Route path="/register" element={<RegisterPage />} />
+        </Routes>
+      </div>
+    </div>
+  );
+};
 
 export default App
